fixed the CORS issue which may cause tainted canvas error
when loading the remote images we might encounter the tainted canvas error, I added the cross origin policy to overcome this situation. note that the remote server need to send proper access control headers in order this to work which means the remote server should allow using their images publicly or you can write a proxy to send your headers which is not legally right.